Initial commit with fixed-width instructions.
This commit has no references to Chunk from the instructions. This
means that OP_CONSTANT doesn't know it's actual value when it
prints, because that's stored in the constant array in chunk. To
fix this, I will need to move responsibility for disassembling
instructions out of OpCode, maybe to Chunk.